+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - update UglifyJS to 3.0.28
 - remove Angular ngInject processing
 - add unsafe_math compress option
+- add parse options
 
 Dropped options: